Added react-aria-components.  This will improve the look and feel of some of the menu and selection items.
Added a /test page that is used to test the new react-aria-components.

Added the Vanilla CSS start kit from the web page: https://react-spectrum.adobe.com/react-aria/Menu.html#starter-kits . This brings in the CSS files for the components that are in the react-aria-components.

Added some of the defined colors in the style.css file.  This redefines the react-aria-components colors from blue to green.  We may need to change these colors again.